requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.3.0
python-dotenv>=1.0.1
openai>=1.11.0
python-dateutil>=2.9.0
# Required by Pydantic for validating email addresses
email-validator>=2.1.0
# Note: the prototype uses a JSON file for persistence instead of a database.
# If you integrate Supabase or other services, add their client libraries here.